requirements.txt
faker
kafka-python
